+++ lib/etudiants/etudiant.h
@@ -4,9 +4,9 @@
 #include "../classes/classe.h"
 
 typedef struct Date{
-	unsigned short int jj;
-	unsigned short int mm;
-	unsigned int aaaa;
+	int jj;
+	int mm;
+	int aaaa;
 } Date;
 
 typedef struct Etudiant{
@@ -18,15 +18,15 @@ typedef struct Etudiant{
 	int code_classe;
 } Etudiant;
 
-Etudiant new_etudiant(unsigned int, char *, char *, char *, Date);
-Date new_date(unsigned short int, unsigned short int, unsigned int);
+Etudiant new_etudiant(int, char *, char *, char *, Date);
+Date new_date(int, int, int);
 
 int get_last_numero_etudiant();
 Etudiant find_etudiant_with_email(char *);
-int find_etudiant_from_classe(unsigned int, Etudiant *);
+int find_etudiant_from_classe(int, Etudiant *);
 
 void save_etudiant(Etudiant *et);
-void update_etudiant(Etudiant , unsigned short int);
+void update_etudiant(Etudiant , int);
 void print_etudiant(Etudiant);
 
 #endif
